+++ src/Authentication/KerberosAuth.php
@@ -23,7 +23,6 @@ use Bolt\protocol\V5_3;
 use Laudis\Neo4j\Contracts\AuthenticateInterface;
 use Laudis\Neo4j\Exception\Neo4jException;
 use Psr\Http\Message\RequestInterface;
-use RuntimeException;
 
 use function sprintf;
 
@@ -43,15 +42,7 @@ final class KerberosAuth implements AuthenticateInterface, Stringable
         private string $token
     ) {}
 
-    /**
-     * @psalm-mutation-free
-     */
-    public function authenticateHttp(RequestInterface $request, string $userAgent): RequestInterface
-    {
-        throw new RuntimeException('Cannot authenticate http requests with Kerberos, use bolt instead.');
-    }
-
-    public function authenticateBolt(V4_4|V5|V5_1|V5_2|V5_3 $bolt, string $userAgent): array
+    public function authenticate(V4_4|V5|V5_1|V5_2|V5_3 $bolt, string $userAgent): array
     {
         $response = $bolt->hello(Auth::kerberos($this->token, $userAgent));
         if ($response->getSignature() === Response::SIGNATURE_FAILURE) {
