+++ src/Bolt/BoltDriver.php
@@ -83,7 +83,7 @@ final class BoltDriver implements DriverInterface
         $config ??= SessionConfiguration::default();
         try {
             GeneratorHelper::getReturnFromGenerator($this->pool->acquire($config));
-        } catch (Throwable $e) {
+        } catch (Throwable) {
             return false;
         }
