+++ src/Contracts/AuthenticateInterface.php
@@ -23,18 +23,11 @@ use Psr\Http\Message\RequestInterface;
 interface AuthenticateInterface
 {
     /**
-     * @psalm-mutation-free
-     *
-     * Authenticates a RequestInterface with the provided configuration Uri and userAgent.
-     */
-    public function authenticateHttp(RequestInterface $request, string $userAgent): RequestInterface;
-
-    /**
      * Authenticates a Bolt connection with the provided configuration Uri and userAgent.
      *
      * @return array{server: string, connection_id: string, hints: list}
      */
-    public function authenticateBolt(V4_4|V5|V5_1|V5_2|V5_3 $bolt, string $userAgent): array;
+    public function authenticate(V4_4|V5|V5_1|V5_2|V5_3 $bolt, string $userAgent): array;
 
     /**
      * Returns a string representation of the authentication.
